+++ src/main/java/org/realityforge/bazel/depgen/record/DependencyCollector.java
@@ -60,7 +60,6 @@ final class DependencyCollector implements DependencyVisitor {
         final List<String> urls =
                 metadata.getUrls(artifact, repositories, _record.getAuthenticationContexts(), _callback);
 
-        final List<String> jsAssets;
         final String sourceSha256;
         final List<String> sourceUrls;
         final String sourcesFilename = artifact.getProperty(Constants.SOURCE_ARTIFACT_FILENAME, null);
@@ -72,11 +71,9 @@ final class DependencyCollector implements DependencyVisitor {
             sourceSha256 = metadata.getSha256(sourcesArtifact.getClassifier(), sourcesArtifact.getFile());
             sourceUrls =
                     metadata.getUrls(sourcesArtifact, repositories, _record.getAuthenticationContexts(), _callback);
-            jsAssets = metadata.getJsAssets(sourcesFile);
         } else {
             sourceSha256 = null;
             sourceUrls = null;
-            jsAssets = null;
         }
         final String externalAnnotationSha256;
         final List<String> externalAnnotationUrls;
@@ -105,8 +102,7 @@ final class DependencyCollector implements DependencyVisitor {
                 sourceUrls,
                 externalAnnotationSha256,
                 externalAnnotationUrls,
-                processors,
-                jsAssets);
+                processors);
     }
 
     @Override